pyproject.toml
[project]
name = "maven-distributed-training"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "modal>=1.2.1",
    "nbdistributed",
    "accelerate>=0.34.2",
    "datasets>=3.0.0",
    "tensorboard>=2.17.1",
    "torch==2.4.1",
    "torchaudio==2.4.1",
    "torchvision==0.19.1",
    "transformers>=4.44.2",
    "torch-tb-profiler>=0.4.3",
    "pyyaml>=6.0.3",
    "tensorboard-plugin-profile>=2.21.1",
]

[tool.uv.sources]
nbdistributed = { git = "https://github.com/muellerzr/nbdistributed" }
